Remove empty strings from categories
It's conceivable that an empty string is a valid category. The Evo schema does not forbid it.

Unfortunately, Leapfrog does not support importing objects when the category columns have both an empty string category and NaN values.

The workaround is to remove the empty string from the categories, which results in empty string values being published as NaN.