Windows batch script that runs the duf-converter. (#129)
## Description

This PR adds a new "samples" folder to the root of the repo. The
intention of this folder is to host various code samples that don't
necessarily fit into the individual converter sub-folders.
The first such code sample is a Windows batch script that publishes a
duf file to an Evo workspace.

## Checklist

- [x] I have read the contributing guide and the code of conduct