Make convert_duf async
The conversion makes use of an async function from the Evo SDK for publishing objects. Making the conversion async avoids having to create or re-enter an event loop for the publishing.

Jupyter notebooks are running their own event loop, and asyncio blocks re-entering or creating a new event loop on the same thread. The library `nest_asyncio` was an attempt to circumvent that restriction, but that was causing problems.

This commit prepares for other converters becoming async.