Update evo-objects package and adjust OMF tests (#125)
The pin on evo-objects in evo-data-converters-common was out of date. Newer code is already making use of 0.3.0+ features, so the pin has been updated to reflect this.

This caused tests to fail in the OMF module, because ObjectReference now requires an HTTPS hub URL rather than None.

This was fixed by adding a standard way to prevent converter functions publishing to Evo with a publish_objects flag, rather than relying on the now unsupported empty hub_url. OMF and other tests were adjusted to use this mechanism and the test URLs provided by evo.common.test_tools to derive the hub_url instead of a blank URL.

In addition to fixing DUF tests broken by the change, there were 4 additional DUF tests that were broken due to recent changes with the CRS handling, so this work was also done:
(1) Ensure that the EPSG codes used in tests are valid
(2) For the tests which were fixed in (1), further adjust their expected results to account for the fact that some of the integer columns are being converted to double on publish

A change was added to prevent collection of DUF tests on non-Windows systems, as this was causing issues running pytest across the whole project on a non-Windows system.

---------

Co-authored-by: Daniel Kinney <97559791+daniel-kinney-seequent@users.noreply.github.com>